requirements.txt
python-dotenv>=1.0.1
pydantic>=2.8.0
langchain>=0.3.0
langgraph>=0.2.39
openai>=1.40.0
requests>=2.32.3
tqdm>=4.66.4
gspread>=6.1.2
google-auth>=2.34.0
chromadb>=0.5.3
pytest>=8.2.0
google-generativeai>=0.7.2
